refactor(security): split security core modules (#0000)
